Implemented the decider logic to return the correct type of CommandType and to Cache the database objects.
Added a script to create a view and renamed the insert data script accordingly.

Added tests to the system to check that the decider was being used correctly.

Found bug in that although CommandType has a TableDirect method it is not actually implemented by SqlClient so currently there is a failing test in relation to this